Refactor database adapters to use a method call
Removed direct calls to the database adapter's methods in dbSync.js and replaced them with a new 'callMethod' utility in DatabaseAdapter. These changes prioritize better code structure and improved method validation. Additionally, it was implemented the 'CAPABILITIES' object in each database adapter sub-class to denote its supported methods, improving the scalability and readability of the code.